+++ libs/impeller2/src/types.rs
@@ -863,6 +863,13 @@ impl From<Timestamp> for hifitime::Epoch {
     }
 }
 
+#[cfg(feature = "hifitime")]
+impl From<hifitime::Epoch> for Timestamp {
+    fn from(epoch: hifitime::Epoch) -> Self {
+        Timestamp((epoch.to_unix_milliseconds() * 1000.0) as i64)
+    }
+}
+
 impl Timestamp {
     pub const EPOCH: Timestamp = Timestamp(0);